+++ src/mantarray_desktop_app/main.py
@@ -272,7 +272,6 @@ def main(
     if shared_values_dict["beta_2_mode"]:
         num_wells = 24
         shared_values_dict["latest_software_version"] = None
-        shared_values_dict["utc_timestamps_of_beginning_of_stimulation"] = [None]
         shared_values_dict["stimulation_running"] = [False] * num_wells
         shared_values_dict["stimulation_info"] = None
         shared_values_dict["stimulator_circuit_statuses"] = {}