+++ tests/main/test_launch.py
@@ -462,7 +462,6 @@ def test_main__full_launch_script_runs_as_expected(fully_running_app_from_main_e
 
     shared_values_dict = app_info["object_access_inside_main"]["values_to_share_to_server"]
     assert shared_values_dict["latest_software_version"] is None
-    assert shared_values_dict["utc_timestamps_of_beginning_of_stimulation"] == [None]
     assert shared_values_dict["stimulation_running"] == [False] * expected_num_wells
     assert shared_values_dict["stimulation_info"] is None
     assert shared_values_dict["stimulator_circuit_statuses"] == {}